hesh/requirements.txt
appdirs==1.4.4
argh==0.26.2
arrow==0.13.1
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
autograd==1.3
autoscraper==1.1.12
backcall==0.2.0
backports.weakref==1.0.post1
bkcharts==0.2
black==19.10b0
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
bs4==0.0.1
cachetools==4.1.1
certifi==2020.12.5
colorama==0.4.4
comtypes==1.1.9
contextlib2==0.6.0.post1
converter==1.0.0
cssselect==1.1.0
cycler==0.10.0
cytoolz==0.11.0
decorator==4.4.2
delayed==0.11.0b1
Delorean==1.0.0
Deprecated==1.2.12
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
ffmpeg==1.4
ffmpeg-python==0.2.0
ffprobe==0.5
Flask-SQLAlchemy==2.4.4
flatbuffers==1.12
future==0.18.2
gast==0.2.2
gmpy2==2.0.8
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
HeapDict==1.0.1
hiredis==2.0.0
httplib2==0.18.1
humanize==3.1.0
imageio-ffmpeg==0.4.2
imbalanced-learn==0.7.0
inflection==0.5.1
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.0
json5==0.9.5
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterthemes==0.20.0
Keras==2.3.1
Keras-Preprocessing==1.1.2
keras-tuner==1.0.1
lesscpy==0.13.0
lightgbm==3.2.1
llvmlite==0.36.0
locket==0.2.1
mccabe==0.6.1
missingno==0.4.2
moviepy==1.0.3
mpmath==1.2.1
multipledispatch==0.6.0
mypy-extensions==0.4.3
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
pafy==0.5.5
parso==0.8.1
pathspec==0.7.0
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pickle-mixin==1.0.2
pickleshare==0.7.5
pkginfo==1.7.0
ply==3.11
proglog==0.1.9
prompt-toolkit==3.0.14
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycrypto==2.6.1
pycurl==7.43.0.6
pydub==0.24.1
Pygments==2.7.4
pynndescent==0.5.2
pyreadline==2.1
pytest==6.2.3
python-dateutil==2.8.1
pytube==10.4.1
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.2
QtPy==1.9.0
redis==3.5.3
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.18.1
scikit-learn==0.24.2
scikit-lego==0.6.6
scipy==1.4.1
simplegeneric==0.8.1
six==1.15.0
sklearn==0.0
sklego==0.0
SpeechRecognition==3.8.1
statsmodels==0.12.0
tables==3.6.1
tabulate==0.8.9
tensorboard==2.1.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.9.4
terminaltables==3.1.0
text-unidecode==1.3
Theano==1.0.4
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
umap-learn==0.5.1
unicodecsv==0.14.1
uritemplate==3.0.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
win-unicode-console==0.5
wincertstore==0.2
wordcloud==1.8.1
wrapt==1.12.1
xgboost==1.3.1
xlwings==0.23.0
xlwt==1.3.0
xmltodict==0.12.0
yellowbrick==1.3.post1
youtube-dl==2021.1.8
zict==2.0.0
zope.event==4.5.0
